app/mcp/providers/file_provider.py
"""
File Provider for MCP - Handles file operations
"""
from typing import Dict, Any, List, Optional
import os
import hashlib
from pathlib import Path
import aiofiles
import magic

from app.models.document import Document
from app.db.session import AsyncSessionLocal
from app.core.config import settings
from app.services.document_processor import DocumentProcessor
from sqlalchemy import select


class FileProvider:
    """Provider for file-related tools"""
    
    def __init__(self):
        self.document_processor = DocumentProcessor()
    
    async def read_file(
        self,
        document_id: str,
        context: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Read the content of a document

        Args:
            document_id: Document ID
            context: Optional context

        Returns:
            Document content
        """
        # Temporarily disabled to prevent transaction corruption
        # TODO: Fix with proper dependency injection
        raise NotImplementedError("Database operations temporarily disabled to prevent transaction corruption")
    
    async def extract_text(
        self,
        document_id: str,
        page_range: Optional[List[int]] = None,
        context: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Extract text from a document

        Args:
            document_id: Document ID
            page_range: Optional page range [start, end]
            context: Optional context

        Returns:
            Extracted text
        """
        # Temporarily disabled to prevent transaction corruption
        # TODO: Fix with proper dependency injection
        raise NotImplementedError("Database operations temporarily disabled to prevent transaction corruption")
    
    async def save_file(
        self,
        filename: str,
        content: bytes,
        file_type: str,
        user_id: int,
        context: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Save a new file

        Args:
            filename: File name
            content: File content
            file_type: File type/extension
            user_id: User ID
            context: Optional context

        Returns:
            Saved file information
        """
        # Temporarily disabled to prevent transaction corruption
        # TODO: Fix with proper dependency injection
        raise NotImplementedError("Database operations temporarily disabled to prevent transaction corruption")
    
    async def delete_file(
        self,
        document_id: str,
        context: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Delete a file

        Args:
            document_id: Document ID
            context: Optional context

        Returns:
            Deletion status
        """
        # Temporarily disabled to prevent transaction corruption
        # TODO: Fix with proper dependency injection
        raise NotImplementedError("Database operations temporarily disabled to prevent transaction corruption")
    
    async def get_file_info(
        self,
        document_id: str,
        context: Optional[Dict[str, Any]] = None
    ) -> Dict[str, Any]:
        """
        Get file information without reading content

        Args:
            document_id: Document ID
            context: Optional context

        Returns:
            File information
        """
        # Temporarily disabled to prevent transaction corruption
        # TODO: Fix with proper dependency injection
        raise NotImplementedError("Database operations temporarily disabled to prevent transaction corruption")