requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
asyncpg==0.29.0

# Search & AI
elasticsearch==8.11.0
qdrant-client==1.7.3
llama-index==0.9.13
langchain==0.0.348
haystack-ai==2.0.0b5
sentence-transformers>=3.0.0
openai==1.3.7
ollama==0.1.7

# Document Processing
pypdf==3.17.1
python-docx==1.1.0
openpyxl==3.1.2
python-pptx==0.6.23
beautifulsoup4==4.12.2
pillow==10.1.0
pytesseract==0.3.10
python-magic==0.4.27

# Caching & Queue
redis==5.0.1
celery==5.3.4

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-sqlalchemy==0.42b0
opentelemetry-exporter-otlp==1.21.0
prometheus-client==0.19.0

# Security & Compliance
cryptography==41.0.7
pycryptodome==3.19.0
bleach==6.1.0
pyotp==2.9.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2
aiosqlite==0.19.0

# Utilities
pydantic==2.11.7
pydantic-settings==2.10.1
email-validator==2.1.0
python-dateutil==2.8.2
pytz==2023.3
aiofiles==23.2.1
tenacity==8.2.3

# WebSocket & Real-time
websockets==12.0
python-socketio==5.10.0

# System Monitoring
psutil==5.9.6